+++ flask-tileserver.env
@@ -1,2 +1,2 @@
-export FLASK_APP=tileserver/application/__init__.py
+export FLASK_APP=localtileserver/application/__init__.py
 export FLASK_ENV=development
