+++ localtileserver/application/static/js/cesium.js
