+++ localtileserver/application/static/js/geojsControls.js
