+++ localtileserver/application/static/styles/style.css
