+++ localtileserver/application/templates/_include/geojsControls.html
