+++ localtileserver/application/templates/cesiumViewer.html
