+++ localtileserver/application/templates/geojsViewer.html
