+++ localtileserver/data/bahamas_rgb.tif
