+++ localtileserver/data/frmt_wms_arcgis_mapserver_tms.xml
