+++ localtileserver/data/frmt_wms_bluemarble_s3_tms.xml
