+++ localtileserver/data/frmt_wms_virtualearth.xml
